- Updated `AdoptionFormController` with OpenAPI annotations and fixed class name. - Refactored `FormFieldController` to fix endpoint mappings and include proper HTTP methods. - Enhanced DTOs (`AdoptionFormCreateDTO`, `AdoptionFormDTO`) with additional documentation and constructors. - Updated `FormFieldCreateDTO` and `FormFieldDTO` to better reflect field attributes and added validation. - Improved `AdoptionFormMapper` and `FormFieldMapper` with new methods and better documentation. - Modified `AdoptionForm` and `FormField` entities to include schema annotations and removed obsolete comments. - Updated `AdoptionFormRepository` with repository documentation. - Made various improvements to `AdoptionFormService`.
