+++ src/main/java/org/mindera/fur/code/mapper/adoptionMapper/AdoptionFormMapper.java
@@ -2,8 +2,8 @@ package org.mindera.fur.code.mapper.adoptionMapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
-import org.mindera.fur.code.dto.formsDTO.AdoptionFormCreateDTO;
-import org.mindera.fur.code.dto.formsDTO.AdoptionFormDTO;
+import org.mindera.fur.code.dto.forms.AdoptionFormCreateDTO;
+import org.mindera.fur.code.dto.forms.AdoptionFormDTO;
 import org.mindera.fur.code.model.form.AdoptionForm;
 
 import java.util.List;
