+++ src/test/java/org/mindera/fur/code/service/AdoptionFormIntegrationTest.java
@@ -3,9 +3,9 @@ package org.mindera.fur.code.service;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.junit.jupiter.api.*;
-import org.mindera.fur.code.dto.formsDTO.AdoptionFormCreateDTO;
-import org.mindera.fur.code.dto.formsDTO.AdoptionFormDTO;
-import org.mindera.fur.code.dto.formsDTO.FormFieldDTO;
+import org.mindera.fur.code.dto.forms.AdoptionFormCreateDTO;
+import org.mindera.fur.code.dto.forms.AdoptionFormDTO;
+import org.mindera.fur.code.dto.forms.FormFieldDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
